+++ tests/IntegrationTest/ObjectToObjectMetadataFactoryTest.php
@@ -36,4 +36,29 @@ class ObjectToObjectMetadataFactoryTest extends AbstractFrameworkTest
         $this->assertEquals(ObjectWithScalarPropertiesDto::class, $metadata->getTargetClass());
         $this->assertCount(4, $metadata->getPropertyMappings());
     }
+
+    public function testProxyResolving(): void
+    {
+        $factory = $this->get('rekalogika.mapper.object_to_object_metadata_factory');
+        $this->assertInstanceOf(ObjectToObjectMetadataFactoryInterface::class, $factory);
+
+        // @phpstan-ignore-next-line
+        eval(<<<'PHP'
+namespace Proxies\__CG__\Rekalogika\Mapper\Tests\Fixtures\Scalar;
+
+class ObjectWithScalarProperties extends \Rekalogika\Mapper\Tests\Fixtures\Scalar\ObjectWithScalarProperties {
+}
+PHP);
+
+        $metadata = $factory->createObjectToObjectMetadata(
+            // @phpstan-ignore-next-line
+            'Proxies\\__CG__\\Rekalogika\\Mapper\\Tests\\Fixtures\\Scalar\\ObjectWithScalarProperties',
+            ObjectWithScalarPropertiesDto::class,
+            Context::create()
+        );
+
+        $this->assertEquals(ObjectWithScalarProperties::class, $metadata->getSourceClass());
+        $this->assertEquals(ObjectWithScalarPropertiesDto::class, $metadata->getTargetClass());
+        $this->assertCount(4, $metadata->getPropertyMappings());
+    }
 }